refactor: refactored js to use bundles
